Moved Network classes for new Response classes
