+++ app/src/main/java/org/asdfgamer/sunriseClock/network/request/DeconzRequestTestConn.java
@@ -1,12 +1,14 @@
-package org.asdfgamer.sunriseClock.network;
+package org.asdfgamer.sunriseClock.network.request;
 
 import android.net.Uri;
 
 import com.android.volley.Response;
 
+import org.asdfgamer.sunriseClock.network.DeconzApiEndpoints;
+import org.asdfgamer.sunriseClock.network.response.IResponseListenerJSONObject;
 import org.json.JSONObject;
 
-public class DeconzRequestTestConn extends DeconzRequest implements  IResponseListenerJSONObject {
+public class DeconzRequestTestConn extends DeconzRequest implements IResponseListenerJSONObject {
 
     public DeconzRequestTestConn(Uri baseUrl, String apiKey) {
         super(baseUrl, apiKey);
