+++ app/src/main/java/org/asdfgamer/sunriseClock/network/response/IResponseListenerJSONArray.java
@@ -1,10 +1,10 @@
-package org.asdfgamer.sunriseClock.network;
+package org.asdfgamer.sunriseClock.network.response;
 
 import com.android.volley.Response;
 
 import org.json.JSONArray;
 
-interface IResponseListenerJSONArray {
+public interface IResponseListenerJSONArray {
 
     /* Should fire  the request towards deconz. Uses own success and error listeners. */
     void fire(Response.Listener<JSONArray> customListenerSuccess, Response.ErrorListener customListenerError);
