+++ app/src/main/java/org/asdfgamer/sunriseClock/network/response/IResponseListenerJSONObject.java
@@ -1,10 +1,10 @@
-package org.asdfgamer.sunriseClock.network;
+package org.asdfgamer.sunriseClock.network.response;
 
 import com.android.volley.Response;
 
 import org.json.JSONObject;
 
-interface IResponseListenerJSONObject {
+public interface IResponseListenerJSONObject {
 
     /* Should fire  the request towards deconz. Uses own success and error listeners. */
     void fire(Response.Listener<JSONObject> customListenerSuccess, Response.ErrorListener customListenerError);
